Move `Clockinterface` to SDK
